Added a test function to create a mocked `AuthorizationStrategy`
This fixes #1038. It also tries to restructure the testfolders a little bit.
