+++ cli/commands/service/orchestrator/orchestrator_test.go
@@ -34,7 +34,7 @@ import (
 
 	"clouditor.io/clouditor/internal/testdata"
 	"clouditor.io/clouditor/internal/testutil/clitest"
-	"clouditor.io/clouditor/internal/testutil/orchestratortest"
+	"clouditor.io/clouditor/internal/testutil/servicetest/orchestratortest"
 	"clouditor.io/clouditor/service"
 
 	"clouditor.io/clouditor/api/assessment"
