+++ service/evidence/evidence_store_test.go
@@ -36,8 +36,9 @@ import (
 	"testing"
 
 	"clouditor.io/clouditor/api"
-	"clouditor.io/clouditor/internal/testutil/evidencetest"
-	"clouditor.io/clouditor/internal/testutil/orchestratortest"
+	"clouditor.io/clouditor/internal/testutil/servicetest"
+	"clouditor.io/clouditor/internal/testutil/servicetest/evidencetest"
+	"clouditor.io/clouditor/internal/testutil/servicetest/orchestratortest"
 	"clouditor.io/clouditor/internal/util"
 
 	"clouditor.io/clouditor/api/evidence"
@@ -315,7 +316,7 @@ func TestService_ListEvidences(t *testing.T) {
 		{
 			name: "Successful Filter Of Evidences (with allowed cloud service)",
 			fields: fields{
-				authz: newAuthorizationStrategy([]string{evidencetest.MockEvidence1.CloudServiceId}),
+				authz: servicetest.NewAuthorizationStrategy(false, []string{evidencetest.MockEvidence1.CloudServiceId}),
 				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
 					assert.NoError(t, s.Create(&evidencetest.MockEvidence1))
 					assert.NoError(t, s.Create(&evidencetest.MockEvidence2))
@@ -414,7 +415,7 @@ func TestService_ListEvidences(t *testing.T) {
 		{
 			name: "Permission denied (cloud service id not allowed)",
 			fields: fields{
-				authz: newAuthorizationStrategy([]string{testdata.MockCloudServiceID}), // allow only MockCloudServiceID
+				authz: servicetest.NewAuthorizationStrategy(false, []string{testdata.MockCloudServiceID}), // allow only MockCloudServiceID
 			},
 			args: args{
 				in0: context.TODO(),
@@ -900,23 +901,4 @@ func TestService_GetEvidence(t *testing.T) {
 			tt.want(t, gotRes)
 		})
 	}
-}
-
-func newAuthorizationStrategy(cloudServiceIDs []string) service.AuthorizationStrategy {
-	return &AuthorizationStrategyMock{
-		cloudServiceIDs: cloudServiceIDs,
-	}
-}
-
-type AuthorizationStrategyMock struct {
-	cloudServiceIDs []string
-}
-
-func (*AuthorizationStrategyMock) CheckAccess(_ context.Context, _ service.RequestType, _ api.CloudServiceRequest) bool {
-	// Not needed (yet)
-	panic("implement me")
-}
-
-func (s *AuthorizationStrategyMock) AllowedCloudServices(_ context.Context) (all bool, IDs []string) {
-	return false, s.cloudServiceIDs
 }
